requirements.txt
python-telegram-bot==22.6
yt-dlp==2026.2.21
aiofiles==25.1.0
python-dotenv==1.2.1
aiohttp==3.13.3
